+++ CHANGELOG.md
@@ -14,6 +14,7 @@ without compromising stable operations.
 
 ## In progress: 0.63.0
 
+- add `bbox` and `geometry` to exported STAC items ([eu-cdse/openeo-cdse-infra#418](https://github.com/eu-cdse/openeo-cdse-infra/issues/418))
 
 ## 0.62.0